*** empty log message ***

+++ src/sccz80/float.h
@@ -4,4 +4,4 @@ extern void fltadd(unsigned char *, unsigned char *);
 extern void fltmult(unsigned char *, unsigned char *);
 extern void pack(unsigned char *, unsigned char *);
 extern void unpack(unsigned char *, unsigned char *);
-extern void norm(long, long, long, long, unsigned char *, long);
+extern void norm(unsigned long, unsigned long, unsigned long, int, unsigned char *, int);
